infra/scripts/agent_scripts/requirements.txt
aiohttp==3.13.4
azure-identity==1.25.2
azure-ai-projects==2.0.0b3
